requirements.txt
flask==2.2.3
python-dotenv==1.0.0
openai==0.28.1
gunicorn==20.1.0
werkzeug==2.2.3